catching more missing references and outdated names
